[algorithms][envelope] restructure file structure for the envelope algorithm
